+++ CHANGES.md
@@ -1,7 +1,10 @@
 ## 3.4.0 -- UNRELEASED
 
-- In composed strings, you may now supply :align with values :left, :right, or :center instead of pad (:right, :left, :both); support for :pad may be removed in the future
-- In `clj-commons.format.table`
+- `clj-commons.ansi`:
+    - In composed strings, you may now supply :align with values :left, :right, or :center instead of pad (:right, :left, :both); support for :pad may be removed in the future
+    - Fonts may now include `double-underlined`
+    - Fonts may now be crossed-out or uncrossed-out (though this is not widely supported) 
+- `clj-commons.format.table`
     - New `miniminal-style` for table output that uses only spaces to separate columns
     - New :title-align and :align keys for columns to be used instead of :title-pad and :pad (support for which may be removed in the future)
     - Table styles now include a :divider? key which, if true, enables the divider between the title line and the first row of data (previously, the divider was not optional)